Module digital_signature::common::ecdsa

This crate wraps libsecp256k1. The message is a str and we use Sha256 for hashing.

Example

Creating a signature on a message is simple.

First, we need to initialise the crate and generate the key-pair which include the public key and the private key.

use digital_signature::common::ecdsa::*;
fn main() {
let mut common_parameter = initialize();
let (public_key, private_key) = key_gen(&mut common_parameter);
}

We can now use the private key to sign our message str.

let message = "hello world";
let (signature, recovery_id) = sign(&message, &private_key).unwrap();

To verify that the signature is valid for our message, we use the public key.

assert!(verify(&message, &signature, &public_key));

Finally, we can recover the public key using the message and the output of the signing function.

assert_eq!(recover(&message, &signature, &recovery_id).unwrap(), public_key);

Structs

CommonParameters

Error

PrivateKey

PublicKey

RecoveryId

Signature

SignatureError

Functions

initialize

key_gen

recover

sha256hash

sign

verify